fixed 2022 のデータを追加（仮）

+++ .eslintrc.js
@@ -13,6 +13,6 @@ module.exports = {
   parser: 'babel-eslint',
   rules: {
     'comma-dangle': ['error', 'always-multiline'],
-    qoutes: ['error', 'double'],
+    qoutes: ['error', 'single'],
   },
 }
